requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
deprecation==2.1.0
exceptiongroup==1.1.3
GeoAlchemy2==0.14.1
gotrue==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.1.0
postgrest==0.16.4
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
realtime==1.0.4
requests==2.31.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.20
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
trio==0.22.2
trio-websocket==0.10.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
webdriver-manager==4.0.0
websockets==12.0
wsproto==1.2.0
